feat: Fixed serialization issue with rain levels in hourly weather forecast
- The level value is received as Float, then the original AemetRainData
value field failed when try to parse to Int.
- Implemented variant AemetRainProbability to separate two parse values.
- Updated affected data models and services.

+++ project/IFCD37WeatherForecast/app/application/src/main/java/com/visualstudioex3/application/ports/input/weather/forecast/models/HourlyWeatherForecastData.kt
@@ -44,7 +44,7 @@ data class HourlyWeatherForecastDataDetail(
      *
      * @return Returns a value in mm.
      */
-    val rain: Int,
+    val rain: Float,
 
     /**
      * Wind chill.
